[refactor] 핸들 컨트롤 분리
- handleController 생성 및 핸들 컨트롤 코드 공통화
- canvasView 접근방식 수정 (Core를 통한 접근)
